CON-359: extract implementation of FinalityDetectorBySingleSweep to a new file
